refactor:divide tests into packages
