+++ src/test/java/chess/model/board/CoordinateTest.java
@@ -1,4 +1,4 @@
-package chess.model;
+package chess.model.board;
 
 import chess.model.board.Coordinate;
 import org.junit.jupiter.api.Test;
