+++ src/test/java/chess/model/dao/ChessDAOTest.java
@@ -1,4 +1,4 @@
-package chess.model;
+package chess.model.dao;
 
 import chess.model.dao.ChessDAO;
 import chess.model.dto.BoardDTO;
